pyproject.toml
[project]
name = "mondiscog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.3",
    "dotenv>=0.9.9",
    "pillow>=9.3.0",
    "python3-discogs-client>=2.8",
]
